Create openvpnfix
This is the script written to stop and start the shill service. It needs to be run from the terminal shell in crosh, not via the native crostini debian stretch container. The crosh shell is the one you get when you type "ctrl+alt+t" in chromeos or at least on my pixelbook.